api/src/requirements/lambda.txt
Flask==1.0.2
flask-restx==0.2.0
requests==2.22.0
python-jose[cryptography]==3.1.0
schema==0.7.2
simplejson==3.16.0
elasticsearch==7.8.0
requests-aws4auth==1.0
cachecontrol[filecache]==0.12.6
